Extract code from test to source and setup first real unit test
