+++ java-frontend/src/test/files/JavaFrontend.java
@@ -6,28 +6,8 @@ class Base {
 
 class MyClass extends Base {
 
-  /*
-  public static String empty(int a) {
-    return "" + a;
-  }
-
-  private static class MyInnerClass {
-
-    private void MyMethod() {}
-
-  }
-
-  public Integer foo(int a, int b, int c, int d, int e, int f, int g) {
-    String v = empty(42);
-    for (int i = a; b < c; d++) {
-      e;
-    }
-    f;
-  }
-  */
-
-  public String myConst(String a) {
-    a = "hehe";
+  public String constLiteral(String a) {
+    return "foo";
   }
 
 }
